+++ fabric/src/main/java/com/iamkaf/mochila/fabric/datagen/ModLanguageProvider.java
@@ -43,6 +43,8 @@ public class ModLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add("key.category.mochila.mochila", "Mochila");
         translationBuilder.add("key.mochila.open_backpack", "Open Backpack");
         translationBuilder.add("key.mochila.open_ender_backpack", "Open Ender Backpack");
+        translationBuilder.add("jei.mochila.category.backpack_coloring", "Backpack Coloring");
+        translationBuilder.add("jei.mochila.category.backpack_upgrading", "Backpack Upgrading");
         translationBuilder.add("tag.item.mochila.backpacks", "Backpacks");
         translationBuilder.add("tag.item.mochila.leather_backpacks", "Leather Backpacks");
         translationBuilder.add("tag.item.mochila.iron_backpacks", "Iron Backpacks");